ipv4: Boolean(opts.ipv4 ? opts.ipv4 === "true" : true),         ipv6: Boolean(opts.ipv6 ? opts.ipv6 === "true" : true),         tailscale: Boolean(opts.tailscale ? opts.tailscale === "true" : true),         public: Boolean(opts.public ? opts.public === "true" : true),

+++ run_ddns_update_once.ts
@@ -77,13 +77,13 @@ export async function run_ddns_update_once(
         }));
         console.log(
             await Promise.all([
-                client.CreateDNSRecord(
+                需要添加的记录内容.length && client.CreateDNSRecord(
                     需要添加的记录内容,
                 ),
-                client.DeleteDNSRecord(
+                需要删除的记录ID.length && client.DeleteDNSRecord(
                     需要删除的记录ID,
                 ),
-            ]),
+            ].filter(Boolean)),
         );
         console.log("更新完成", { 需要删除的记录ID, 需要添加的记录内容 });
     }
